requirements.txt
Flask
google-genai==1.7.0
python-docx
chromadb